pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monarch-mcp"
dynamic = ["version"]
description = "Model Context Protocol (MCP) server for Monarch Money personal finance platform"
readme = "README.md"
license = "MIT"
authors = [
    { name = "vargahis" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.0",
    "monarchmoneycommunity~=1.3.0",
    "keyring>=24.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "asyncio>=3.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pylint>=3.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/vargahis/monarch-mcp"
Repository = "https://github.com/vargahis/monarch-mcp"
Issues = "https://github.com/vargahis/monarch-mcp/issues"

[project.scripts]
monarch-mcp = "monarch_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
fallback_version = "0.0.0"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]